Finish Win32 polling. Improve LOG and ASSERT naming.

+++ Bhazel/src/Bhazel/Platform/GLFW/GlfwInput.h
@@ -9,8 +9,8 @@ namespace BZ {
         bool isKeyPressedImpl(int keycode) override;
 
         bool isMouseButtonPressedImpl(int button) override;
-        std::pair<float, float> getMousePositionImpl() override;
-        float getMouseXImpl() override;
-        float getMouseYImpl() override;
+        std::pair<int, int> getMousePositionImpl() override;
+        int getMouseXImpl() override;
+        int getMouseYImpl() override;
     };
 }
